Move serde code into separate file
